examples/tweet_intro/requirements.txt
oauth2
Pillow
requests